# swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "version": "1.0.0",
      "title": "Store Manager",
      "description": "API docs for Store Manager.",
      "contact": {
         "name": "Emmanuel Raji",
         "email": "manuelraji15@gmail.com"
      }
   },
   "servers": [
      {
         "url": "https://onlinestoremanager.herokuapp.com/api/v1"
      }
   ],
   "components": {
      "securitySchemes": {
         "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      }
   },
   "paths": {
...
      "/products": {
         "get": {
            "tags": [
               "Products"
            ],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "summary": "Returns all products in store",
            "responses": {
               "200": {
                  "description": "Successfully fetched products",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "type": "object",
                              "required": [
                                 "products"
                              ]
                           }
                        }
                     }
                  }
               },
               "403": {
                  "description": "you are not authorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "500": {
                  "description": "internal server error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               }
            }
         },
         "post": {
            "tags": [
               "Products"
            ],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "summary": "Lets an admin add new product to store",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "required": [
                           "product"
                        ],
                        "properties": {
                           "productName": {
                              "type": "string"
                           },
                           "price": {
                              "type": "integer"
                           },
                           "productQuantity": {
                              "type": "integer"
                           },
                           "minimumInventory": {
                              "type": "integer"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "201": {
                  "description": "Successfully created product",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "message": {
                                 "type": "string"
                              },
                              "success": {
                                 "type": "boolean"
                              }
                           }
                        }
                     }
                  }
               },
               "403": {
                  "description": "you are not authorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "422": {
                  "description": "Invalid input details",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "boolean"
                              }
                           }
                        }
                     }
                  }
               },
               "500": {
                  "description": "internal server error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/products/{productId}": {
         "get": {
            "tags": [
               "Products"
            ],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "summary": "Returns the specified product",
            "parameters": [
               {
                  "name": "productId",
                  "in": "path",
                  "required": true,
                  "description": "  Enter the product id",
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successfully fetched product",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "items": {
                              "type": "object",
                              "properties": {
                                 "product_id": {
                                    "type": "integer"
                                 },
                                 "product_name": {
                                    "type": "string"
                                 },
                                 "product_price": {
                                    "type": "integer"
                                 },
                                 "product_quantity": {
                                    "type": "integer"
                                 },
                                 "minimum_inventory": {
                                    "type": "integer"
                                 }
                              }
                           }
                        }
                     }
                  }
               },
               "403": {
                  "description": "you are not authorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "product not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "success": {
                                 "type": "boolean"
                              },
                              "message": {
                                 "type": "string"
                              }
                           }
                        }
                     }
                  }
               },
               "500": {
                  "description": "internal server error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "tags": [
               "Products"
            ],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "summary": "Lets an admin modify a product",
            "parameters": [
               {
                  "name": "productId",
                  "in": "path",
                  "required": true,
                  "description": "The id of the product to be modified",
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object",
                        "required": [
                           "product"
                        ],
                        "properties": {
                           "productName": {
                              "type": "string"
                           },
                           "price": {
                              "type": "integer"
                           },
                           "productQuantity": {
                              "type": "integer"
                           },
                           "minimumInventory": {
                              "type": "integer"
                           }
                        }
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successfully updated product",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "items": {
                              "type": "object",
                              "properties": {
                                 "message": {
                                    "type": "string"
                                 },
                                 "success": {
                                    "type": "boolean"
                                 }
                              }
                           }
                        }
                     }
                  }
               },
               "403": {
                  "description": "you are not authorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "invalid product",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "success": {
                                 "type": "boolean"
                              },
                              "message": {
                                 "type": "string"
                              }
                           }
                        }
                     }
                  }
               },
               "422": {
                  "description": "Invalid input details",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "boolean"
                              }
                           }
                        }
                     }
                  }
               },
               "500": {
                  "description": "internal server error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "tags": [
               "Products"
            ],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "summary": "Lets an admin delete a product",
            "parameters": [
               {
                  "name": "productId",
                  "in": "path",
                  "required": true,
                  "description": "The ID of the product to delete",
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successfully deleted product",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "items": {
                              "type": "object",
                              "properties": {
                                 "message": {
                                    "type": "string"
                                 },
                                 "code": {
                                    "type": "boolean"
                                 }
                              }
                           }
                        }
                     }
                  }
               },
               "403": {
                  "description": "you are not authorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "product not fount",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "code": {
                                 "type": "boolean"
                              },
                              "error": {
                                 "type": "string"
                              }
                           }
                        }
                     }
                  }
               },
               "500": {
                  "description": "internal server error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               }
            }
         }
      },
...
   }
}